Initial sd_card_usci commit
Added assembly code to perform CRC7 calculations, this is to setup
sending out commands to the SD card. Tweeked the makefile to include a
few more linker warning, also added a debug symbol for the asm code to
use. Added the overall block diagram wiring image too.

The CRC7 subroutine needs to be tweeked in order to check for
transmission errors when reading back responses from the SD card.